android-crmtab/app/src/main/java/fr/pds/isintheair/crmtab/helper/CheckInternetConnexion.java
package fr.pds.isintheair.crmtab.helper;

import android.content.Context;
import android.net.ConnectivityManager;
import android.net.NetworkInfo;

/**
 * Created by tlacouque on 30/12/2015.
 *
 */
public class CheckInternetConnexion {

    /**
     * Called to know if the user have an  internet connexion
     * @param context
     * @return boolean
     */
    public static boolean isNetworkAvailable(Context context) {
        ConnectivityManager connectivityManager
                = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
        return activeNetworkInfo != null && activeNetworkInfo.isConnected();
    }
}
